create dummy data, adjusted types
