+++ src/widgets/appointment/CalendarMonth.tsx
@@ -3,8 +3,8 @@ import { Box, Center, Divider, Text } from 'native-base';
 import CalendarWeek from './CalendarWeek';
 
 type MonthProps = {
-    year: string;
-    month: string;
+    year: number;
+    month: number;
     appointments: any;
 };
 
@@ -15,8 +15,8 @@ const CalendarMonth: React.FC<MonthProps> = ({ year, month, appointments }) => {
                 <Text>{`${Info.months('long', { locale: 'de-DE' })[Number(month) - 1]} ${year}`}</Text>
             </Center>
             <Divider my={3} />
-            {Object.entries(appointments).map((week) => {
-                return <CalendarWeek appointments={week[1]} />;
+            {Object.entries(appointments).map((weekAppointments) => {
+                return <CalendarWeek appointments={weekAppointments[1]} />;
             })}
         </Box>
     );
