+++ src/widgets/appointment/CalendarYear.tsx
@@ -2,7 +2,7 @@ import { Box } from 'native-base';
 import CalendarMonth from './CalendarMonth';
 
 type YearProps = {
-    year: string;
+    year: number;
     appointments: any;
 };
 
@@ -10,7 +10,7 @@ const CalendarYear: React.FC<YearProps> = ({ year, appointments }) => {
     return (
         <Box>
             {Object.entries(appointments).map((month) => {
-                return <CalendarMonth year={year} month={month[0]} appointments={month[1]} />;
+                return <CalendarMonth year={Number(year)} month={Number(month[0])} appointments={month[1]} />;
             })}
         </Box>
     );
